# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Domakin API",
        "description": "API documentation for Domakin platform",
        "contact": {
            "email": "support@domakin.nl"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "https://orange.domakin.nl",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/renting/searching/create": {
            "post": {
                "tags": [
                    "Renting"
                ],
                "summary": "Submit a property search request",
                "operationId": "58683eeb9312d7c47d63c0e2f24deae7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname",
                                    "phone",
                                    "email",
                                    "people",
                                    "type",
                                    "moveIn",
                                    "period",
                                    "registration",
                                    "budget",
                                    "city",
                                    "terms"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+31 6 12345678"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "people": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "room"
                                    },
                                    "moveIn": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-01"
                                    },
                                    "period": {
                                        "type": "string",
                                        "example": "12 months"
                                    },
                                    "registration": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "budget": {
                                        "type": "number",
                                        "example": 1200
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Amsterdam"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "referralCode": {
                                        "type": "string"
                                    },
                                    "letter": {
                                        "description": "Optional motivational letter",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "terms": {
                                        "description": "JSON string with terms",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Search request submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/viewing/details/{id}": {
            "get": {
                "tags": [
                    "Viewing"
                ],
                "summary": "Get viewing details",
                "operationId": "84814bb61e42dc90c91a8067c6cd9ee1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Viewing ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/viewing/create": {
            "post": {
                "tags": [
                    "Viewing"
                ],
                "summary": "Create a viewing appointment",
                "operationId": "d61ebda8e31388e0f5f3dd3f5ba543c2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "property_id",
                                    "name",
                                    "email",
                                    "phone",
                                    "date",
                                    "time"
                                ],
                                "properties": {
                                    "property_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+31 6 12345678"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-10"
                                    },
                                    "time": {
                                        "type": "string",
                                        "example": "14:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Viewing created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}